Provide a caching scoreboard. This majorly improves the scalability of DOMjudge; displaying is now O(1) time in stead of O(tp). A non-scientific showed that the new scoreboard loads in <1 sec while the old took over 10 seconds!
How does this work?
We have a two cache tables, scoreboard_{public,jury} with rows (cid,team,problem,
submissions,totaltime,correct,penalty). Upon submission, the row (cid,team,problem)
is recalculated (entirely, but only that row) and either:
- inserted into both tables (when submittime < lastscoreupdate), or
- inserted only into the jury table (when submittime > lastscoreupdate).
A row is also updated on a rejudge, and a function is added which allows for a
total recalculation in case of problems.

[actual implementation differs on details]

Also added:
- a new 'admin' menu, with functions only for the administrator (as opposed to
the rest of the jury)
- more comments



git-svn-id: https://secure.a-eskwadraat.nl/svn/domjudge/trunk@685 caf94f87-7bd6-0310-88dc-dee8202e2269