New required activatetime field for a contest: defines the moment the team/public displays start to display this contest, but no submissions accepted yet. Solves the problem of testsession data still displayed upto the second before a contest starts (Closes: #1825555).
Adds a new field to contests. This is a required field but can be equal
to starttime. It's required because this allows us to keep the code
that determines the current contest simple, and with one simple query
instead of multiple. Since that function is used very often, it seems
a worthwhile tradeoff to make it a required field; especially since I
can imagine that one would in most cases prefer to 'activate' before
'start'ing a contest.



git-svn-id: https://secure.a-eskwadraat.nl/svn/domjudge/trunk@1946 caf94f87-7bd6-0310-88dc-dee8202e2269