On judgehost startup, check for any unfinished judging in this judgehost's name.
When a judgehost crashes, it leaves unfinished judgings in the judging table
(that show as 'judging'). When a judgehost starts up, it knows it's not
currently judging those, so it can conclude that these need to be released
again to the queue before this judgehost starts its judging process.

This eliminates the need to manually press "RESTART judging" for crashed
judgings.