Add support for printing from within DOMjudge.
Using system printing is still the prefered way, as this is less
cumbersome and allows teams to print from within any editor.
For situations where this is not feasible, one can enable the
function in DOMjudge.

The exact command to be used for printing can be configured
in lib/www/printing.php. The code is contained so that if
you don't use it, it will not get in the way.

Loosely inspired by a patch from  Constantin Jucovschi from
Jacobs University Bremen.