Address small memory leak in submit client upon cURL failure.
The curl cleanup functions weren't called in the case an error
condition was met. Also, in case errnum is 0, errdescr should
also be freed in lib.error.c.

These memory leaks have very minor impact because all memory
is reclaimed when the program terminates, and the submit
client runs separately for each submission.