Add problem statements for example problems to the database.
This may be a bit unwieldy in the sql file, given that each
problem statement is about 250 kB. However, I think it's outweighed
by the advantage of having also the problem statement feature
demonstrated through the example data.

One way to improve it, may be by making the pdf's somehow smaller,
not requiring 250 kB for one page of problem statement.