Tighten permissions on judgedaemon output files.
Set a umask in judgedaemon, so any created files will not be world readable
by default. Open up the permissions of the workdir at the start of a run,
so the domjudge-run user can access it. Revoke those permissions afterward.

This improves security for the case where chroot is disabled, because
submissions can no longer read reference testcase output files, or those of
other submissions. However, we still recommend against running with chroot
disabled, as there are othe potential surprises possible.