runguard: Unshare namespaces when using cgroups.
unshare() a number of namespaces, including NEWNET. This means that e.g.
networking will not be available inside runguard, therefore obviating the
need for special measures to encapsulate it.

I'm currently piggybacking on the USE_CGROUPS macro to not make things
more complex, while namespaces are stricly not cgroups, they're a closely
related technology, so this seems acceptable for now.