Move logic of testcase selection, results priority to API.
The judgehost now basically loops until it gets a submission out of
the api, for that submission loops as long as it gets testcases for
it and judges them, then restarts the outer loop.

This moves all handling of results priority, lazy evaluation &c to
the server side, making that argubly a bit complex, but the judgehost
many times simpler.