Rename user.authtoken to user.password; drop obsolete team.authtoken.
Stop overloading the password field with multiple tokens, because
this hinders us when presenting an appropriate UI for the field.
IP address is already stored in the ip_address field.

Changes behaviour for LDAP, which will now just match user.username,
which shouldn't be problematic. Drops unused column team.authtoken.