Improve user management by allowing to set password and adding links.
User passwords may now be (re)set from the user edit page, not on
a separate genpasswds page. The latter is now reserved to only
setting bulk passwords; which is now limited to only those users
with a team role, as we probably do not want to do bulk password
setting for admins or judgehosts.