Move updating of last_login to do_login function.
Otherwise it would be called on every page load which has shown to
be problematic in the past. The drawback to this change is that it
now only updates this info for auth methods that actually perform
a login operation. There is no real way around that unless we keep
a session for any login type, which is planned.