Fix Python scripts by setting a non-existent HOME.
Remove old fix in dj_make_ubuntu_chroot since this was
error prone, e.g. if you add run users after creating the chroot.

If HOME is not set in environment, python{2,3} tries to lookup the home
in the /etc/passwd for the current user id. If there is no user with
the current id in /etc/passwd the python versions in wheezy do bail out
with run error on every submission. In jessie, python only spits out a
warning but does not include the same set of default modules (e.g.
exit() does not longer work without importing sys).