Add full blown rejudgings that do not get applied immediately.
No change necessary in submission pickup mechanism. But instead of
invalidating the old valid judging for that submission and creating a
new valid judging, we create an invalid judging that references the
old valid judging (which is still valid). The new (invalid) judging
becomes valid if the rejudging is manually applied.

A submission can be part of only up to one rejudging at a time which
has to be applied or canceled to delete the association.

TODO: create a simple interface to create rejudgings based on certain
criteria or/and events.