Make problems by default submittable.
This simplifies setting up new problems and importing jury
solutions for it from a problem zip file. The minimal security
risk that sensitive data is exposed because the problem is now
also visible is too exceptional, especially the admin can nowadays
first add the problem itself before coupling it to a contest.