Add a config check whether the timezone is properly configured.
PHP doesn't automatically infer this from the system, although
this is patched in many Linux distributions, including Debian.
Still, the timezone is not read as a location string, but
indirectly via the offset, which can cause issues, e.g. by
misinterpreting Europe/London as UTC (the former switches to DST).