Generate a syslog LOG_ERR message when compare script fails and append compare output. Closes #179.
The point is that when a compare script error occurs, (currently)
the judging doesn't finish and the compare output will not be shown
in the jury interface. Thus there is no point in writing it to
the judgemessage.txt file. Instead it should be written to the logs
and console with a high priority.