Merge rankcache and scorecache {public,jury} variants into single tables.
This gains us performance as it's much more efficient to update several
more fields in one INSERT or UPDATE than to perform two of those
operations separately. Also we can avoid calculating the rank twice in
a row.

The split was done originally when there were different MySQL users
with different privileges. This is no longer the case.

We don't bother to migrate cached data. Please refresh the caches
after you upgrade.