Be nicer to the kernel FS cache
Inform the kernel that we no longer need any of the files after a
judgement is complete. We do this by evicting the files in workdir from
the kernel fs cache as soon as judging for the submission is complete.

This fixes a bug discovered at the 2016WF where java submissions will
run slower when the kernel filesystem cache is nearly full. By flushing
the cache of all the files for each submissions we can avoid this.