Add ordinal to problems API endpoint according to https://clics.ecs.baylor.edu/index.php/Draft_CCS_REST_interface#GET_baseurl.2Fproblems
For now, we set the ordinal based on the short_name in the contest. We
might want to add an internal field in the future, if we would want to
support shadowing behind a primary with a different implementation.

Ordinal starts with 0, and doesn't skip numbers. This is currently
unspecified in the official spec.