Generate Symfony parameters.yml file from dbpasswords.secret.
Note that dbpasswords.secret might not be available yet during
'make dist'; in that case we generate a stub file that's a simple
copy of the parameters.yml.dist template. This will be replaced
later during installation with a generated file that contains the
correct database passwords.

Also drop sqlite config example from template file.