Use symfony to handle login/logout
Only difference against legacy code is that we don't support being
logged in to the api from an existing session, you have to basic auth
against it.

We could enable this, but it also means the reverse, basic auth against
the api would log you in to the web interface. Not sure that's
desirable(To do this, just add context: domjudge to the security.yml
firewalls for api and main, and remove stateless:true from api)