Split get_image_thumb_type() into two functions and fix error handling.
Fail more graceously when the GD library is not installed or for
whatever reason fails, so that we can still process other data, e.g.
when importing testcases.

We leave error handling to the calling function, by returning FALSE on
failure and the error in the pass-by-reference variable $error. Now the
calling function can decide how to deal with this, e.g. die or not.

Closes #273.