Remove eventlog() from clarification create transaction.
The eventlog() function calls the REST API that in turn does a DB
query to get the clarification with ID just generated. But that one
was still in transaction, and hence fails to return anything.
To work around this, we only log both to event (and for consistency)
to the auditlog after the clarification is created. We also update
the team_unread table afterwards, but that's not crucial.
Closes #286.