Drop first_to_solve from API scoreboard: it's not needed and wrong.
Note that the error is a minor and subtle one: if previous to the
first correct submission on a problem there are still pending
submissions, then these can still become first to solve. Thus we
can only assign first_to_solve after all these are judged.

Finally, if we implement this, we also have to be careful of how to
handle the frozen scoreboard, to not leak information.