Only require entry point for java/kt/py2/py3.
This should be done properly by making it configurable in the database,
but it should do its job for NWERC.

If JS is enabled, we hide the entry_point input field. As soon as a
language gets selected (either manually or by autodetect), the entry
point input field is displayed again if necessary. The text is changed
to reflect that we look for a class or a file name depending on the
language.