Add transactions and reorder queries to improve API consistency.
In the API we should enforce referential integrity. This means
for example that a new judgement should never be exposed before
the submission is. We can fix this by wrapping the eventlog()
call inside an SQL transaction.

The submission ignore is a bit a special case. Here we only
perform the eventlog earlier. This case is not well supported
yet anyways, since we should delete/(re)create dependent
judgements and runs.