Fix time comparisons: always check with $start <= $time < $end.
Note that a time interval starts at the beginning of the
second (or whatever the resolution used) of the indicated time.
Thus, if a contest starts at 12:30:00, then the whole second
of times starting from 12:30:00.00000... is inside the contest.
On the other hand, if the contest ends at 15:00:00, then
14:59:59.999999 is still in the contest, but 15:00:00 is not
anymore.