Improve show_flags/show_affiliations settings.
If one disables flags, likely they do not want them to show up in
other places like the team info pages in the public interface.
When show_flags is disabled, flags are not shown anywhere. If
show_affiliations is disabled, these do not show other than in
the admin interface under 'affiliations'. Also the scoreboard
filter does not offer to filter on these dimensions if disabled.