Only expose too-late submissions in API when requested by ID.
In 167af6fdd408e6e4e072132 we stopped exposing too-late submissions,
but this caused eventlog to throw an error at submit time, since
the submission would not be available in the API for inserting as
event.

Now we only expose too-late submissions if queried for with a
specific ID. This doesn't cause a security risk since the same
role restrictions are in place and too-late submissions were not
especially sensitive to begin with.

(cherry picked from commit ed21fda022f1d8aabf0006fd72dbcf1534146d9b)