Send judgement.endtime to the event-feed when lazy-eval is off.
We were not sending an update event for the judgement end time
when lazy evaluation is off. The specification does not clearly
specify when it should be send, but currently we fix it by sending
when all testcases are judged. Note that this doesn't handle
verification required well. We need to revisit this.

(cherry picked from commit 369a239915a056ac4c1914ea4cb7a93ae8e17667)