Add defense against reltive paths and symlinks in zipfiles.
This is only a problem when actually extracting on the filesystem,
as unzip does. Add "-j" option to junk any paths when extracting,
because we already did that when building the zip file.

Whether a zipfile contains symlinks is hard to check and we cannot
tell unzip to ignore them. Check with "unzip -Z" to see if any are
listed. ZipArchive is oblivious to symlinks.