Remove static_scoreboard script
This didn't work anymore because it calls public/index.php on the
command line, missing some setup. Did not want to fix that, rather,
just have people use `curl` to fetch it via an URL - this reduces
the number of code paths (rather than increasing them to fix this
issue). Considered replacing static_scoreboard with a script doing
the `curl` request but it became too trivial.