Clarify "rejudge on same judgehost" restriction.
Disabling this means that a submission cannot be rejudged on
the judgehost if it was already judged on it before. The main
use case is if you want to rejudge a set of submissions various
times, so that each submission is judged on each judgehost once.

(cherry picked from commit b83ce7552d33de964e6ff548ef12daa0bf8e2590)

# Conflicts:
#	www/jury/judgehost_restriction.php