Don't break if input/output are missing for a testcase.
We now create an internal error with a proper error message instead and
disable the problem.

I still don't know how to get in that state through normal operations (I
did by manipulating the database). If we know how it can happen, we
should create an error message earlier (not while judging).