Allow batch updates for single judging runs.
Note that it doesn't work yet because:
A new entity was found through the relationship
'DOMJudgeBundle\Entity\JudgingRunWithOutput#judging' that was not
configured to cascade persist operations for entity:
DOMJudgeBundle\Entity\Judging@00000000357318ca0000000078db4686. To solve
this issue: Either explicitly call EntityManager#persist() on this
unknown entity or configure cascade persist this association in the
mapping for example @ManyToOne(..,cascade={"persist"}). If you cannot
find out which entity causes the problem implement
'DOMJudgeBundle\Entity\Judging#__toString()' to get a clue.