Reintroduce AJAX updates for menu bar alerts.
Now implemented with jQuery which is a lot cleaner.

This was not ported from the old interface yet. It only updates the
numbers; sending desktop notifications when a number changes, or
doing a page reload (is that needed?) is still missing.

Moved all setting of these numbers to the JS, no longer also do it
in the HTML, to reduce code duplication and logic in the twig templates.