Make it possible to specify sorting data for a column.
This allows one to display a formatted value (e.g. 2 kB) but sort
that column on a sortable value (e.g. 2048).

Also fix titles as empty title on <a> will disable a title set in <td>.
Also fix flag display when DOMjudge is not served from the root URL
and add alt fallback.