Try umount a little bit harder.
If umounting fails, let's try once more before we give up. Also, if it
fails the second time (even with force) again, print why this is the
case to make debugging easier.

The concrete in which this showed up on my machine was a Gnome daemon
called tracker that seems to have started indexing the mounted chroot.