Remove balloons from default access list.
It basically means giving balloons user by default access to any
page not specifying otherwise. This is not desirable.

The structural fix is that we migrate things to Symfony. This
also alleviates the original proble that the change tried to fix.
For pages tht still need this method, $REQUIRED_ROLES needs to
be declared global specifically.

This reverts 62d90da26c60bdef179c45c8b5d938dc19dd08af.