Fix import/export of results. Now requires php-intl extension.
Change default locale to just "en" because it's fine for us and yields
better error if PHP intl extension is missing. Add php-intl to requirements.
We could work around it when not present, but that seems not worth it at
the moment.

Closes: #475
(cherry picked from commit 703cf073c57d157710ec85c3f5b2a068ebd9b153)