Move jury clarification pages to Symfony.
This is not perfect and inherits some of the shortcomings/complexities
of the original system and data model. It however seems to work at least
as good and some of the bugs in the old system have been fixed. Further
refactoring is possible but should at least be easier in the current
system.