Add option to write out metadata from runpipe.
Also change semantics a little bit:
Previously, we returned the first failing exit code of any of the two
commands as exit code of the runpipe command. Now we special case it a
bit since runpipe is used in a single place and the jury validator is
always the first and sudo/runguard the second command: the jury
validator returns 42/43 depending on the verdict and we return that as
the runpipe exit code.