Implement combined run and compare scripts for interactive problems.
There's still a difference in judging compared to problemtools which
needs further discussion:
What's the correct verdict if a submission is both TLE and WA and the WA
is determined before the time limit? For non-interactive problems, the
answer is a clear TLE, but problemtools returns WA for those cases.

Part of #465.