Move config checker to Symfony.
It's now available both via the API and via the web interface.

The code is not the most beautiful, sometimes literally taken
from the old code. But that can be improved upon. Some tests
in the old checker that are now obsolete have not been skipped.
Also not the integrity checker since that is no longer very
relevant and requires quite some complexity (already in the old
interface) so would be nice to get rid of that.

The check runs in 4,5 sec on my dev environemnt, or about 800 ms
if you leave out the submissions. We could investigate how useful
we think these checks are and see if we can cut down on runtime.
In any case it's quite faster than the old one.

Kept the old page and infrastructure around for now.