Only show problem as First To Solve if there are no earlier pending submissions
The time in scorecache is now set to the time of the earliest pending
or correct submission, while before it was only set for the earliest
correct one.

We can use this to determine if we are ready to award the first-to-solve
prize yet. We update the summary of the minimum time for any cell which
has a correct or pending submission. We only show that summary line
when a dark green cell has actually been awarded.