Store first to solve information in the scorecache.
This moves all logic to determine what constitutes a first to solve
to one place, calculation of the score row. When refreshing a row,
and the judgement is correct, determine whether there are any earlier
submissions for this contest, problem, sort order that are either
correct or still pending. If this is the case, this submission is
not (yet) the first to solve.

Advantage is that this extra complexity and time is spent in updating
the cache. Every display of the scoreboard or team row is now even
simpler than before so remains fast. Drawback is that a change of
sortorder information now also requires recalculating of the cache.