drop combined_scoreboard tool
It is broken in several ways (strict types, API changes, stuff that
has been removed). Each can probably be fixed but it still largely
integrates with obsolete things so it would more or less mean to
rebuild it. Combined with the lack of evidence that someone actually
uses this, it seems better to remove it now which allows to remove
its remaining dependencies. And when there's demand, make something
that works with our current technologies.