Implement IP authentication autologin
In the old implementation, when IP authentication was enabled, you still needed to click a button to actually log in. This commit makes a configuration option available `ip_autologin` (default off) which immediately logs in the user.

Furthermore, after login; you can now visit `/login` to log in with a different user.