Add country validation
The old version of the `TeamAffiliation` did not do any validation on the `$country` property, this resulted in a database error when random text was entered, to prevent this, I implemented a validator which checks whether the entered value is either empty or a valid ISO3166-1 alpha-3 value. A phpleague library provides the list of the countrycodes; this might be overkill. Another option is to simply add them (to Utils perhaps) as a const.