Implemented max-width for team-, affiliate-names (#542)
Implemented max-width for team-, affiliate-names.

Added a configuration option to set the max-width for the affiliation and team names. When set to 0, it does not apply a maximum width. Some JS adds a class when the contents are too wide, which adds a slot that houses the gradient, this gradient overlaps with the rightmost 30% of the field.

Fixes #431