Initialize static events when clients connect to the event feed and when referenced events are missing when inserting new events.
This was the other reason we needed the eventdaemon.
Note that if a referenced event is present, but does not contain up to date data, we will not insert it when inserting a new event.
This is something that would be quite slow to fix and the workaround is to just add a new client to listen to the event feed for a short amount of time, so we can live with this.