CheckConfig: remove loop over all submissions.
This is what makes ConfigCheck really slow, without it it's fast on
a reasonably sized contest. It seems we never had a use for the checks
this did. If there's interest in a specific check, we can add it
with a query for that condition instead of looping over all submissions
in the database.