Clean up registration form.
We used to display two error messages relating to the username needing to be only alphanumeric characters, one also allowing - and _.
This was because we have a default validation on the username property of a user and we added one ourselves as well.
Also took the time to clean up the registration form code.
Thanks Mark van Helvoort for reporting the bug!