Fixes the ugly toggle switch at the config page
This was caused due to the fact that the default "Off" is narrower than "Yes", doing a quick reset of this css property fixes this issue.

Tried it in all my browsers (FF, Opera, and Safari) and it did not cause more issues then were already present. As far as I am aware and could find, this is the only place where this library gets used; so the fixed is placed in the config page instead of modifying `bootstrap-toggle.min.js`.