Fix batch judging updates.
The output_limit is specified in kB, the unsent data is in B.
This speeds up judging of a problem with 190 test cases on my machine
from 69s to 51s.

The time per test case is ~275ms. A breakdown:
- time to start testcase_run: ~15ms
- set up program: ~25ms
- run program: ~85ms
  * sudo: ~15ms
  * process exit after unshare: ~50ms
- set up compare script: ~5ms
- run compare script: ~90ms
  * sudo: ~15ms
  * process exit after unshare: ~50ms
- post-processing results in testcase_run: ~35ms
- API POST time (divided by batch size): ~25ms

Possibilities for improvement are:
- unshare less stuff for the compare script, e.g. allow network (-30ms)
- do not change user for compare script, run as judgedaemon user (-15ms)
- reduce inefficiencies in testcase_run (or rewrite in c) (-40ms)
- run a script/binary that does both runguard calls and is called with
  sudo permissions (-15ms)

Doing all of this might reduce the overhead by 100ms, or the whole run
to ~32s.