Reindex JSON elements from API to be able to determine correct element for events. Fixes #621
If you edited a problem that was not part of a contest, the code could crash because it assumed it got exactly the expected number of elements from the API.
Now we first reindex the results by ID, so we can actually pick the right element.
If the right element does not exist, skip it but log a warning at the end.
The warning used to be an exception but that was not very user friendly.

(cherry picked from commit 34d8e0492f7b2e7e8c395689e4170157a7b71985)