Turn Testcase->content and JudgingRun->output in a 'fake' OneToMany relation to make lazy loading work.
Doctrine will always load a OneToOne relation where the foreign key is in the *other* table, since that is the only way for it to know whether the value should be NULL or an actual object (or proxy object).
Since the "other side" contains a lot of data that we do not want to load, we turn it into a OneToMany/ManyToOne relation and just get the first value from the collection.
Our code makes sure there is always only one item in the collection.

This idea is based on https://github.com/doctrine/orm/issues/4389#issuecomment-162367781