Always display a flash message for skipped submissions in rejudgings.
This makes it harder to miss why a submissions is not included in a
rejudging. For example, if you earlier did do a rejudging of problem X
which is still active (i.e. you didn't apply nor cancel it) and then
later (perhaps after fixing the testdata) did do the same rejudging it
would only rejudge the submissions that are not already part of the
first, still active, rejudging.

See #641.