Fix PHPUnit.
The Symfony PHPUnit bridge expects the vendor directory to be in the root of the project,
see https://github.com/symfony/phpunit-bridge/blob/master/bin/simple-phpunit.php#L70-L76.

However, we don't do that so the bridge will install the latest version of itself (which is now 5.0) instead of using the existing version.
This breaks stuff if Symfony releases a new major version (like they just did).