Fix cancelling a rejudging, (partially) fixes #656.
The original code was migrated to Symfony in 04bc0bb0a34b114a4,
which changed a left join on judging to a join. Likely the old
code was also wrong because it forced the judging to exist.

Then da7f59e486a25a53d049968 attempted to fix this, but likely
only worked in case of applying the rejudging, not cancelling it,
because in that case there can still be submissions without
judgings for the given rejudging.

This roughly reverts to the old pre-Symfony code, except that
the clause "j.rejudgingid = ?" is now part of the left join,
since we also want to get the submissions with judgings.

Also get rid of the groupBy's since they're not needed, and
j.result wasn't used either.