Fix edge case bug in first to solve calculation.
Submission.judgehost IS NULL is not a correct way to detect
whether a submission has no (valid) judging associated to it.

Note that a submission has exactly zero or one valid judgings,
so make that explicit in the LEFT JOIN query. Then, we can
check that no valid judging exists with j.judgingid IS NULL.