Rename team.externalid to team.icpcid. Fixes #655
For other entries we use the external ID as the 'primary' ID as given by either the primary CCS or another system.
However, for teams we always use team.id for this and team.externalid was a (second) external ID, always representing the ID from the ICPC CMS.
This gave a lot of confusion (as can also be seen in the commit, since I have fixed a few places we used it incorrectly).
So I believe it is better to rename it to something else to not have any confusion.