Document how to use REMOTE_USER for authentication
This is a very generic solution for authenticating existing users.
Many (Apache) modules can handle a full authentication flow and
deliver an authenticated user, thereby instantly supporting a range
of methods for DOMjudge.

It does not (yet) support auto creation of users. That might need
more custom work per method, e.g. where to find the full name
or email address? But it's definitely better than what we have now.
Surely we can improve on it further.

Since this does provide a way to authenticate with OpenID Connect,
I think it's time to finally clean up the openid_ configuation
values which have not been working for many releases now.