Fix submit help to work when base url is incorrect/non responsive.
There's a catch 22 in that the help text incorporates elements from
the API but when the API would be completely unavailable, submit
would fail fast thereby also not proviing the user with help to
e.g. specify another URL.

This makes the api information requests only warn when they cannot
get the information, so processing can continue.
(The submit POST will still fail hard when it cannot reach the API.)

Also move handling of --version to above the API requests since they
are not needed at all to produce this output.

Closes: #707