Use exceptions to bubble up and handle issues at the right level.
The choice of whether to warn or error out should be handled
by the caller function who knows whether the issue is recoverable
or not. That's the read* functions in this case.

Currently doAPIsubmit() still throws errors, but that's ok-ish,
since it's the final call that does determine overall success
or failure.