Fix permissions on webapp/var/* by using group setsid bit.
Use a small script that we can call from the Makefile and when
running dj_setup_database. Since in many situations root will
have default access to mysql, it's likely this command is run
as root. That will screw up permissions of cache files too, so
try to fix them here too.

Closes #650