Fix link to documentation from the admin index page.
The previous link only worked on the maintainer-install of DOMjudge.
Make it so we link to the right location at the regular install,
which also drops the ugly `build/` from the url path. And add a
symlink on the maintainer-install so the docs work there also.

Closes #698

(cherry picked from commit 98857da37bfe3c1ee9f92fbb9031ffc631fdeb21)