Allow to "delete" testcases even with associated judgings.
This only marks testcases as deleted but doesn't actually delete them.
Since the association with the problem is removed they will not be taken
into account with future (re)judgings of the problem.

All testcases that are ranked after the deleted one are moved up one
rank.

In a follow up change we can actually delete testcases which don't have
any associated judgings.

With commit d873b7c we added a warning to submissions which were judged
either in a different testcase order or with a now deleted testcase.

We may want to offer to rejudge all of those cases together with a
single click but I didn't want to bundle to much in this commit.

Fixes #189.