Drop smart renaming for Java public classes.
This creates problems when rejudging or forwarding to
shadow CCS's because the entry point detected doesn't
match the source file: it was renamed during compilation
but this is not stored in the submission.

Also sync Java and Kotlin compile scripts to use DetectMain,
likely an oversight in b5f983a0925a095d1f6ce3aa7536f.

Closes #735