Add simple option to switch between multiple doc versions.
This assumes that the docs for the multiple versions are arranged in a
structure like:
  <SERVER_URL>/docs/manual/<VERSION>/index.html
e.g.
  https://www.domjudge.org/docs/manual/7.2/
  https://www.domjudge.org/docs/manual/7.3/
  https://www.domjudge.org/docs/manual/master/
and that
  <SERVER_URL>/docs/manual/versions.json
contains a version list in JSON format, e.g.
  ["7.2", "7.3", "master"]
.

If the versions.json file is not readable, then the version div remains
unchanged and is not transformed to a select dropdown.