Also terminate user session on a role change of the user.
When a user's password was changed or the Enabled status flipped,
the user is automatically logged out. This was however not the
case for a change in roles. Meaning that e.g. revoking Jury rights
from a user may not be effective as long as they do not log out.
This fixes that so any change in user roles also causes the
session to be invalidated.