Always install python3-yaml, never python-yaml
It is required by gen_conf_ref.py, which uses `python3` in the shebang
line and so will always run on Python 3, never Python 2.

This is in contrast with the python-sphinx package, which may need to
be Python 2 to match the rst2pdf package.

Also move python3-yaml to the end to emphasise that python-sphinx and
rst2pdf are related.