Process escape sequences in TSV import
Together with commit e42f594d3 (Escape newlines and other reserved
characters in TSV export, 2020-05-18), this allows arbitrary data to
be exported and imported losslessly via TSV.

This implementation uses `stripcslashes`, which also supports other
escape sequences that are not used in the export format. Those escape
sequences can't appear in the exported data, so there's no harm.

(It's tempting to use `str_replace` instead of `stripcslashes`, and
mirror `toTsvField()`, but it's tricky because `str_replace` performs
each substitution in order on the entire string. For example `\\t`
should be decoded as `\t`, but the naive approaches either process it
twice and give `<TAB>`, or disregard precedence and give `\<TAB>`.
Here is a correct approach:

    return str_replace(
        ["\\\\", "\\t", "\\n", "\\r", "\\\\e"],
        ["\\\\e", "\t", "\n", "\r", "\\"],
        explode("\t", rtrim($line, "\r\n"))
    );

However, `stripcslashes` is simpler.)