Change balloon daemon to use the API instead of direct database connection.
* Allows the daemon to be run on any system with API access;
* Can remove duplicated logic and just follow the processed output of the API.
* API also allows other implementations of balloon notification.

The BALLOON_CMD has been moved into the script since it really only pertains
to this script and is therefore only of interest to those actually using it.

API client access and credential reading is kept as simple as possible. There
might be a small market for an API client library when more scripts will use
this API directly.