Timout jobs to trigger trap condition
Currently we timeout jobs on GitLab to retry these faster. When a job is
stuck on a logical error for example the server is always in a HTTP500
state we would keep on repeating this problem without ever getting the
logs. This will timeout before as a soft limit to trigger the trap and
when this fails the hard timout would terminate the whole job.