Make sure progname doesn't contain '%' when used in formatting string.
Escape all '%' by a double "%%" so that when progname contains
any '%', these will just be reproduced verbatim when using
`buffer` as formatting string later, instead of leading to
a potential format string attack. Note that that risk is small
though as the attacker must be able to manipulate the name
of the program being run.

This issue was detected by https://lgtm.com/