Added an abstract testing framework
Instead of replicating code for generact add/edit/delete pages this is
now handled in a generic class and then filled with the values in the
inheriting classes.

Constructor is used to fill some dynamic values,

Tried to search most values via the framework instead of assuming them
to be consistent over the years. Based on the found values the URLs are generated.

Currently missing:
- checking if the modal approach does work
- checking if the delete modal and delete page
  look correct

Move the adding and editing to a follow-up PR.

The LanguageController endpoint was broken, but is fixed in via another commit.